requirements.txt
aiogram~=3.6.0
aiohttp~=3.9.0
aiohttp-socks~=0.8.4
aiohttp_cors~=0.7.0
aiofiles~=23.2.1
aiojobs==1.2.0
tenacity~=8.2.2
environs~=9.3.4
async_lru~=2.0.2
loguru~=0.6.0
humanize~=4.8.0
pillow~=10.1.0
curl_cffi~=0.5.10
pydantic~=2.7.1

pixivpy3~=3.7.2
beautifulsoup4~=4.10.0
Pybooru~=4.2.2

# DATABASE
# Local:
tinydb~=4.5.2
# Firebase:
firebase-admin~=6.1.0
# Mongo:
motor~=2.5.1
dnspython~=2.1.0
# Cloudant:
cloudant~=2.15.0

# STORAGE
# Mega:
git+https://github.com/y-young/mega.py@6015359#egg=mega.py
# Google Drive:
PyDrive2~=1.15.1
# S3
minio~=7.1.16
